+++ src/kernel/service_cpio.c
@@ -358,7 +358,7 @@ static unsigned int protocol_seek(struct task_descriptor *descriptor, unsigned i
 
 }
 
-static unsigned long protocol_map(struct task_descriptor *descriptor)
+static unsigned int protocol_map(struct task_descriptor *descriptor)
 {
 
     struct cpio_header header;
