+++ src/kernel/x86/mmu.h
@@ -37,6 +37,6 @@ struct mmu_directory
 
 };
 
-void mmu_map(struct mmu_directory *directory, struct mmu_table *table, unsigned long paddress, unsigned long vaddress, unsigned int size, unsigned int tflags, unsigned int pflags);
+void mmu_map(struct mmu_directory *directory, struct mmu_table *table, unsigned int paddress, unsigned int vaddress, unsigned int size, unsigned int tflags, unsigned int pflags);
 void mmu_setdirectory(struct mmu_directory *directory);
 void mmu_setup(void);
