+++ src/mboot/mboot.h
@@ -14,7 +14,7 @@ struct mboot_header_modules
 {
 
     unsigned int count;
-    unsigned long address;
+    unsigned int address;
 
 };
 
@@ -23,7 +23,7 @@ struct mboot_header_elf
 
     unsigned int shcount;
     unsigned int shsize;
-    unsigned long shaddress;
+    unsigned int shaddress;
     unsigned int shstringindex;
 
 };
@@ -32,7 +32,7 @@ struct mboot_header_mmap
 {
 
     unsigned int count;
-    unsigned long address;
+    unsigned int address;
 
 };
 
@@ -40,7 +40,7 @@ struct mboot_header_drives
 {
 
     unsigned int count;
-    unsigned long address;
+    unsigned int address;
 
 };
 
@@ -79,8 +79,8 @@ struct mboot_mmap
 {
 
     unsigned int size;
-    unsigned long address0;
-    unsigned long address1;
+    unsigned int address0;
+    unsigned int address1;
     unsigned int limit0;
     unsigned int limit1;
     unsigned int type;
@@ -90,7 +90,7 @@ struct mboot_mmap
 struct mboot_module
 {
 
-    unsigned long address;
+    unsigned int address;
     unsigned int limit;
     unsigned int name;
     unsigned int reserved;
